feat: removed appwrite and also removed page router
