+++ package.json
@@ -30,7 +30,6 @@
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-toggle-group": "^1.0.4",
     "@radix-ui/react-tooltip": "^1.0.7",
-    "appwrite": "^13.0.1",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "globby": "^14.0.0",
