--- src/app/(auth)/login/page.tsx
@@ -1,7 +1,0 @@
-import LoginPage from '@/pages/loginPage'
-
-const Login = () => {
-  return <LoginPage />
-}
-
-export default Login
