--- src/app/(auth)/signup/page.tsx
@@ -1,9 +1,0 @@
-import dynamic from 'next/dynamic'
-
-const SignUpPage = dynamic(() => import('@/pages/signupPage'), { ssr: false })
-
-const Signup = () => {
-  return <SignUpPage />
-}
-
-export default Signup
