+++ src/components/header/avatarDropDown.tsx
@@ -9,20 +9,13 @@ import {
   DropdownMenuShortcut,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu'
-import appwriteService from '@/appwrite/config'
 import { useEffect, useState } from 'react'
 import { useRouter } from 'next/navigation'
 
 const DropdownMenuDemo = () => {
   const router = useRouter()
   const [user, setUser] = useState<any>(null)
-  useEffect(() => {
-    const userData = async () => {
-      const data = await appwriteService.getCurrentUser()
-      setUser(data)
-    }
-    userData()
-  }, [])
+
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
@@ -47,16 +40,6 @@ const DropdownMenuDemo = () => {
             <DropdownMenuShortcut>⌘S</DropdownMenuShortcut>
           </DropdownMenuItem>
         </DropdownMenuGroup>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem
-          onClick={async () => {
-            await appwriteService.logout()
-            router.push('/login')
-          }}
-        >
-          <span>Log out</span>
-          <DropdownMenuShortcut>⇧⌘Q</DropdownMenuShortcut>
-        </DropdownMenuItem>
       </DropdownMenuContent>
     </DropdownMenu>
   )
